get using livedata

+++ app/src/main/java/com/tinaciousdesign/covidtoday/data/Country.kt
@@ -3,20 +3,20 @@ package com.tinaciousdesign.covidtoday.data
 import com.tinaciousdesign.covidtoday.networking.CountryInfoResponse
 
 data class Country(
-    val active: Int,
-    val activePerOneMillion: Int,
-    val cases: Int,
-    val casesPerOneMillion: Int,
-    val critical: Int,
-    val criticalPerOneMillion: Int,
-    val deaths: Int,
-    val deathsPerOneMillion: Int,
-    val recovered: Int,
-    val recoveredPerOneMillion: Int,
+    val active: Float,
+    val activePerOneMillion: Float,
+    val cases: Float,
+    val casesPerOneMillion: Float,
+    val critical: Float,
+    val criticalPerOneMillion: Float,
+    val deaths: Float,
+    val deathsPerOneMillion: Float,
+    val recovered: Float,
+    val recoveredPerOneMillion: Float,
     val country: String,
     val countryInfo: CountryInfoResponse,
-    val todayDeaths: Int,
-    val todayCases: Int,
-    val todayRecovered: Int,
+    val todayDeaths: Float,
+    val todayCases: Float,
+    val todayRecovered: Float,
 //    val updated: // Epoch time stamp
 )
